+++ lib/smollm2.dart
@@ -1,6 +1,7 @@
 /// SmolLM2 Library.
 library;
 
+export 'src/chat.dart';
 export 'src/export_smollm2.dart';
 export 'src/quant_type.dart';
 export 'src/smollm2.dart';